Implement lab1 ex3 MatrixMultiplaction
Define the services in the proto file. Use nested, repeated message
to define the arrays A, B and C. Establish two rpc services for
the functions 'multiplyBlock' and 'addBlock'.

Reuse the GrpcServer.java file from the baeldung example.

Implement the two functions 'multiplyBlock' and 'addBlock' in the
'CalculatorServiceImpl.java' file . Parse the request to arrays
A and B. Execute the mathematical functions. Build a response
message.

Functions 'multiplyBlock' and 'addBlock' in the GrpcClient class
build the request message, call the stub and parse the response
message to array C.

The biggest challenge in this lab exercise was to deal with arrays
in java and translate them to grpc message and back to java arrays.